refactor: ajusta BookRepository e outros para usar ID Long no JpaRepository

+++ cadastro-livros/src/main/java/com/portfolio/cadastro_livros/infrastructure/repository/BookRepository.java
@@ -3,6 +3,6 @@ package com.portfolio.cadastro_livros.infrastructure.repository;
 import com.portfolio.cadastro_livros.infrastructure.entitys.Book;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BookRepository extends JpaRepository<Book, Integer> { // o JPA já possui métodos prontos para o CRUD
+public interface BookRepository extends JpaRepository<Book, Long> { // o JPA já possui métodos prontos para o CRUD
 
 }
